Fix calculation error for average calculation
- Now using UD60x18 correctly to calculate the average and also the two thirds.
- Before i casted values to uint/int where the numbers lost their precision
- Returning 1000 from calculateTwoThirdAverage, because this is a value which
can't be computed as an average. If only one player would guess 1000, the
average would be 1000 but two third of that would be less than 1000. Therefore
this is a safe to assume value, which can't be reached and is treated as the
value, that gets returned, if no participants values have been submitted.